# api/openapi.yaml
openapi: 3.0.3
info:
  title: ZJC 航空公司机票预订系统 API
  description: ZJC 航空公司机票预订系统 REST API 接口文档
  version: 1.0.0
  contact:
    name: API Support
    email: support@zjc-airline.com
servers:
  - url: http://localhost:8080
    description: 本地开发服务器
  - url: http://api.zjc-airline.com
    description: 生产服务器
...
paths:
  /api/users/register:
    post:
      tags:
        - 用户管理
      summary: 用户注册
      description: 新用户注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              name: "张三"
              idNumber: "110101199001011234"
              phone: "13800138001"
              address: "北京市朝阳区建国路88号"
              username: "zhangsan"
              password: "123456"
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
        '400':
          description: 注册失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/login:
    post:
      tags:
        - 用户管理
      summary: 用户登录
      description: 用户登录验证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: "admin"
              password: "123456"
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 登录失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/users/{id}:
    get:
      tags:
        - 用户管理
      summary: 根据ID获取用户
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseUser'
    put:
      tags:
        - 用户管理
      summary: 更新用户信息
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: 更新成功
    delete:
      tags:
        - 用户管理
      summary: 删除用户
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 删除成功
...
  /api/bookings/user/{userId}:
    get:
      tags:
        - 预订管理
      summary: 获取用户的所有预订
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 成功
...
components:
  schemas:
    RegisterRequest:
      type: object
      required:
        - name
        - idNumber
        - phone
        - username
        - password
      properties:
        name:
          type: string
          description: 姓名
        idNumber:
          type: string
          description: 身份证号
        phone:
          type: string
          description: 手机号
        address:
          type: string
          description: 地址
        username:
          type: string
          description: 用户名
        password:
          type: string
          description: 密码
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: 用户名
        password:
          type: string
          description: 密码
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        idNumber:
          type: string
        phone:
          type: string
        address:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    ApiResponseUser:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/User'
...
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        data:
          type: object
          nullable: true